widget: Fix 16-bit overflow limiting scrollable lists to 65535px
Scroll positions and sizes were stored as uint16_t/UWORD throughout the scroll machinery (W_SLIDER pos/numStops, ScrollBarWidget, the ClipRect content offset, and ScrollableListWidget/ScrollableTableWidget).

A list taller than 65535px wrapped, cutting off scrolling partway through the content.

Widen the whole chain to uint32_t, and fix the math that only breaks once it turns unsigned.

Also add an opt-in minimum scrollbar-handle size (ScrollBarWidget::setMinimumSliderSize, with a ScrollableListWidget pass-through) so very tall content can keep the handle grabbable. The default remains a proportional handle.

+++ lib/widget/table.h
@@ -173,8 +173,8 @@ public:
 
 	void setItemSpacing(uint32_t value);
 
-	uint16_t getScrollPosition() const;
-	void setScrollPosition(uint16_t newPosition);
+	uint32_t getScrollPosition() const;
+	void setScrollPosition(uint32_t newPosition);
 
 	bool isUserDraggingColumnHeader() const;
